requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
numpy==1.23.4
pandas==1.5.1
pytz==2022.1
python-dateutil==2.8.2
sqlparse==0.4.2
tqdm==4.64.1
tzdata==2022.1